# backend_school_crm/docs/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact:
    email: support@school-crm.example
    name: School CRM Support
  description: Residual monolith — auth, subscriptions, Click.uz/Telegram payments,
    messaging, reports, dev panel.
  license:
    name: Proprietary
  termsOfService: http://swagger.io/terms/
  title: School CRM API
  version: "1.0"
paths:
...
  /branches:
    get:
      description: Returns all branches belonging to the authenticated admin.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List branches
      tags:
      - branches
    post:
      consumes:
      - application/json
      description: Creates a new branch under the authenticated school owner.
      parameters:
      - description: name, address, phone
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: created branch
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "402":
          description: subscription_limit_reached
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create branch
      tags:
      - branches
  /branches/{id}:
    delete:
      parameters:
      - description: Branch UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: branch deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete branch
      tags:
      - branches
    put:
      consumes:
      - application/json
      parameters:
      - description: Branch UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update branch
      tags:
      - branches
...
  /users:
    get:
      description: Returns all users visible to the authenticated admin.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    description: JWT token — prefix with "Bearer "
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"